requirements.txt
PyYAML==6.0.2
tiktoken==0.9.0
torch==2.6.0
transformers==4.51.3
vllm==0.8.5.post1
regex==2024.11.6
bitsandbytes
peft
accelerate
